backend/src/requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2024.1
